finder fix

+++ src/map_service/finder/finder.cpp
@@ -2,7 +2,7 @@
 
 #include <limits>
 
-std::vector<size_t> PathFinder::FindPath(const size_t &from, const size_t &to, const size_t &size,
+std::vector<int> PathFinder::FindPath(const size_t &from, const size_t &to, const size_t &size,
                                          const std::vector<std::vector<double>> &AdjacencyTable)
 {
     auto distance = std::vector<double>(size);
@@ -51,7 +51,7 @@ std::vector<size_t> PathFinder::FindPath(const size_t &from, const size_t &to, c
         vizited[minindex] = 0;
     } while (minindex < limit_index);
 
-    auto path = std::vector<size_t>(size,-1);
+    auto path = std::vector<int>(size,-1);
     auto end = B;
     path[0] = B + 1;
     int k = 1;
