+++ src/map_service/finder/finder.h
@@ -9,7 +9,7 @@ class PathFinder
   public:
     PathFinder() = default;
 
-    std::vector<size_t> FindPath(const size_t &from, const size_t &to, const size_t &size,
+    std::vector<int> FindPath(const size_t &from, const size_t &to, const size_t &size,
                                  const std::vector<std::vector<double>> &AdjacencyTable);
 
     ~PathFinder() = default;
